Moved the test class to the appropriate file and removed the redundant one
